Add a script to automatically add publications to website (#4)
Add a script which runs via the CI once a month. The script uses OpenAlex to search for all relevant articles. It does this by searching for the keywords "gysela", "gyselax" and "gyselalib" amongst articles published since a given date. The date is 2019-01-01 by default (which corresponds with the most recent article currently indexed), but it is usually equal to the date at which the workflow last ran successfully minus 2 months. It filters the articles found to:
- Remove preprints
- Keep articles with the keyword in the title or abstract
- Keep articles where one of the authors is one of the people in https://github.com/gyselax/gyselax.github.io/tree/main/content/authors

For each of the remaining articles a `cite.bib` file and an `index.md` file are created in an appropriately named sub-folder in https://github.com/gyselax/gyselax.github.io/tree/main/content/publication
if any publications are added then a branch is created with these changes and a PR is opened.

If PRs are merged regularly the same article should not appear in more than 1 PR. Once the PR has been created it can be pruned manually before merging if articles were inappropriately added.